Added: Connection error handling
When possible, if we lose connection we'll clear the scheduled setTime() function and now also re-run handleLogin() when we reconnect.

This fixes the client never updating the bot display name if it reconnects, and also the client completely crashing if the connection fails.

There may still be an unhandled promise on the initial disconnect, but for now it doesn't seem to cause issues.

Fixed some format inconsistencies in index.js too.